add arduino executable
split into multiple crates
add procedural macro to generate static song
split necessary lib structs into base crate

+++ ard-r-sound/src/parse_tree.rs
@@ -77,7 +77,7 @@ pub(crate) enum Octave {
 #[derive(Debug)]
 pub(crate) enum Length {
     Default,
-    Specified { divided: bool, number: u64 },
+    Specified { divided: bool, number: u32 },
 }
 
 impl TryInto<crate::abc::Length> for Length {
